requirements.txt
astroid==2.9.0
attrs==21.2.0
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==5.0.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pylint==2.12.1
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
